Move tile and breadcrumb to common for AB#16182
